requirements.txt
requests>=2.28.1
pandas>=1.5.0
numpy>=1.23.0
streamlit>=1.12.0
plotly>=5.11.0
skyfield>=1.45.0
python-dotenv>=0.21.0
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
pydantic>=1.10.4
apscheduler>=3.10.1
python-dotenv>=0.21.0
